requirements.txt
anyio==4.10.0
asgiref==3.11.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
coverage==7.9.2
crashtest==0.4.1
distlib==0.4.0
Django==5.2.8
djangorestframework==3.16.1
dulwich==0.22.8
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.6.3
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
mccabe==0.7.0
more-itertools==10.8.0
msgpack==1.1.1
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pbs-installer==2025.9.2
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.4.0
pluggy==1.6.0
poetry==2.1.4
poetry-core==2.1.3
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-cov==6.2.1
python-dotenv==1.0.1
python-telegram-bot==21.6
pywin32-ctypes==0.2.3
RapidFuzz==3.14.1
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.9.9.12
types-requests==2.32.4.20250809
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
virtualenv==20.32.0
yt-dlp==2025.12.8
zstandard==0.24.0
